+++ apps/klaviyo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # saleor-app-klaviyo
 
+## 1.13.11
+
+### Patch Changes
+
+- Updated dependencies [00070dc3]
+  - @saleor/apps-shared@1.13.0
+
 ## 1.13.10
 
 ### Patch Changes